backend/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
django-filter
django-extensions
django-debug-toolbar
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0